add state transition table and fix tests
